refactor(decoder): コアから分離
